Move LocalizedRun to models package for better organization
- Move LocalizedRun class from utils/timezone.py to models/run.py
- Update models/__init__.py to export LocalizedRun
- Update timezone.py to import LocalizedRun from models
- Add zoneinfo import to run.py for LocalizedRun functionality
- Better co-location of related model classes
